# curved ec performance