--- a/src/solvers/dgsem_structured/dg_3d.jl
+++ b/src/solvers/dgsem_structured/dg_3d.jl
@@ -204,14 +204,14 @@ end
     u_node = get_node_vars(u, equations, dg, i, j, k, element)
 
     # pull the contravariant vectors in each coordinate direction
-    Ja11_node, Ja12_node, Ja13_node = get_contravariant_vector(1, contravariant_vectors,
-                                                               i, j, k, element)
+    Ja1_node = get_contravariant_vector(1, contravariant_vectors,
+                                        i, j, k, element)
 
-    Ja21_node, Ja22_node, Ja23_node = get_contravariant_vector(2, contravariant_vectors,
-                                                               i, j, k, element)
+    Ja2_node = get_contravariant_vector(2, contravariant_vectors,
+                                        i, j, k, element)
 
-    Ja31_node, Ja32_node, Ja33_node = get_contravariant_vector(3, contravariant_vectors,
-                                                               i, j, k, element)
+    Ja3_node = get_contravariant_vector(3, contravariant_vectors,
+                                        i, j, k, element)
 
     # All diagonal entries of `derivative_split` are zero. Thus, we can skip
     # the computation of the diagonal terms. In addition, we use the symmetry
@@ -222,14 +222,12 @@ end
     for ii in (i+1):nnodes(dg)
       u_node_ii = get_node_vars(u, equations, dg, ii, j, k, element)
       # pull the contravariant vectors and compute the average
-      Ja11_node_ii, Ja12_node_ii, Ja13_node_ii = get_contravariant_vector(1, contravariant_vectors,
-                                                                          ii, j, k, element)
-      Ja11_avg = 0.5 * (Ja11_node + Ja11_node_ii)
-      Ja12_avg = 0.5 * (Ja12_node + Ja12_node_ii)
-      Ja13_avg = 0.5 * (Ja13_node + Ja13_node_ii)
+      Ja1_node_ii = get_contravariant_vector(1, contravariant_vectors,
+                                             ii, j, k, element)
+      Ja1_avg = 0.5 * (Ja1_node + Ja1_node_ii)
       # compute the contravariant sharp flux in the direction of the
       # averaged contravariant vector
-      fluxtilde1 = volume_flux(u_node, u_node_ii, SVector(Ja11_avg, Ja12_avg, Ja13_avg), equations)
+      fluxtilde1 = volume_flux(u_node, u_node_ii, Ja1_avg, equations)
       multiply_add_to_node_vars!(du, alpha * derivative_split[i, ii], fluxtilde1, equations, dg, i,  j, k, element)
       multiply_add_to_node_vars!(du, alpha * derivative_split[ii, i], fluxtilde1, equations, dg, ii, j, k, element)
     end
@@ -238,14 +236,12 @@ end
     for jj in (j+1):nnodes(dg)
       u_node_jj = get_node_vars(u, equations, dg, i, jj, k, element)
       # pull the contravariant vectors and compute the average
-      Ja21_node_jj, Ja22_node_jj, Ja23_node_jj = get_contravariant_vector(2, contravariant_vectors,
-                                                                          i, jj, k, element)
-      Ja21_avg = 0.5 * (Ja21_node + Ja21_node_jj)
-      Ja22_avg = 0.5 * (Ja22_node + Ja22_node_jj)
-      Ja23_avg = 0.5 * (Ja23_node + Ja23_node_jj)
+      Ja2_node_jj = get_contravariant_vector(2, contravariant_vectors,
+                                             i, jj, k, element)
+      Ja2_avg = 0.5 * (Ja2_node + Ja2_node_jj)
       # compute the contravariant sharp flux in the direction of the
       # averaged contravariant vector
-      fluxtilde2 = volume_flux(u_node, u_node_jj, SVector(Ja21_avg, Ja22_avg, Ja23_avg), equations)
+      fluxtilde2 = volume_flux(u_node, u_node_jj, Ja2_avg, equations)
       multiply_add_to_node_vars!(du, alpha * derivative_split[j, jj], fluxtilde2, equations, dg, i, j,  k, element)
       multiply_add_to_node_vars!(du, alpha * derivative_split[jj, j], fluxtilde2, equations, dg, i, jj, k, element)
     end
@@ -254,14 +250,12 @@ end
     for kk in (k+1):nnodes(dg)
       u_node_kk = get_node_vars(u, equations, dg, i, j, kk, element)
       # pull the contravariant vectors and compute the average
-      Ja31_node_kk, Ja32_node_kk, Ja33_node_kk = get_contravariant_vector(3, contravariant_vectors,
-                                                                          i, j, kk, element)
-      Ja31_avg = 0.5 * (Ja31_node + Ja31_node_kk)
-      Ja32_avg = 0.5 * (Ja32_node + Ja32_node_kk)
-      Ja33_avg = 0.5 * (Ja33_node + Ja33_node_kk)
+      Ja3_node_kk = get_contravariant_vector(3, contravariant_vectors,
+                                             i, j, kk, element)
+      Ja3_avg = 0.5 * (Ja3_node + Ja3_node_kk)
       # compute the contravariant sharp flux in the direction of the
       # averaged contravariant vector
-      fluxtilde3 = volume_flux(u_node, u_node_kk, SVector(Ja31_avg, Ja32_avg, Ja33_avg), equations)
+      fluxtilde3 = volume_flux(u_node, u_node_kk, Ja3_avg, equations)
       multiply_add_to_node_vars!(du, alpha * derivative_split[k, kk], fluxtilde3, equations, dg, i, j, k,  element)
       multiply_add_to_node_vars!(du, alpha * derivative_split[kk, k], fluxtilde3, equations, dg, i, j, kk, element)
     end