--- a/src/solvers/dgsem_unstructured/dg_2d.jl
+++ b/src/solvers/dgsem_unstructured/dg_2d.jl
@@ -172,14 +172,13 @@ end
     # x direction
     for ii in (i+1):nnodes(dg)
       u_node_ii = get_node_vars(u, equations, dg, ii, j, element)
-      flux1 = volume_flux(u_node, u_node_ii, 1, equations)
-      flux2 = volume_flux(u_node, u_node_ii, 2, equations)
       # pull the contravariant vectors and compute the average
       Ja11_node_ii, Ja12_node_ii = get_contravariant_vector(1, contravariant_vectors, ii, j, element)
       Ja11_avg = 0.5 * (Ja11_node + Ja11_node_ii)
       Ja12_avg = 0.5 * (Ja12_node + Ja12_node_ii)
-      # compute the contravariant sharp flux
-      fluxtilde1 = Ja11_avg * flux1 + Ja12_avg * flux2
+      # compute the contravariant sharp flux in the direction of the
+      # averaged contravariant vector
+      fluxtilde1 = volume_flux(u_node, u_node_ii, SVector(Ja11_avg, Ja12_avg), equations)
       multiply_add_to_node_vars!(du, alpha * derivative_split[i, ii], fluxtilde1, equations, dg, i,  j, element)
       multiply_add_to_node_vars!(du, alpha * derivative_split[ii, i], fluxtilde1, equations, dg, ii, j, element)
     end
@@ -187,14 +186,13 @@ end
     # y direction
     for jj in (j+1):nnodes(dg)
       u_node_jj = get_node_vars(u, equations, dg, i, jj, element)
-      flux1 = volume_flux(u_node, u_node_jj, 1, equations)
-      flux2 = volume_flux(u_node, u_node_jj, 2, equations)
       # pull the contravariant vectors and compute the average
       Ja21_node_jj, Ja22_node_jj = get_contravariant_vector(2, contravariant_vectors, i, jj, element)
       Ja21_avg = 0.5 * (Ja21_node + Ja21_node_jj)
       Ja22_avg = 0.5 * (Ja22_node + Ja22_node_jj)
-      # compute the contravariant sharp flux
-      fluxtilde2 = Ja21_avg * flux1 + Ja22_avg * flux2
+      # compute the contravariant sharp flux in the direction of the
+      # averaged contravariant vector
+      fluxtilde2 = volume_flux(u_node, u_node_jj, SVector(Ja21_avg, Ja22_avg), equations)
       multiply_add_to_node_vars!(du, alpha * derivative_split[j, jj], fluxtilde2, equations, dg, i, j,  element)
       multiply_add_to_node_vars!(du, alpha * derivative_split[jj, j], fluxtilde2, equations, dg, i, jj, element)
     end