Flux differencing on curved meshes sucked since the resulting EC methods were only 25% faster than Fluxo :wink: To the rescue!

<details>
<summary>

This speeds up the RHS evaluation in `joinpath(examples_dir(), "structured_3d_dgsem", "elixir_euler_ec.jl")` by 1/3 for `flux_ranocha` and 1/4 for `flux_shima_etal`. Thus, our EC methods should be roughly twice as fast as Fluxo on curved meshes.

</summary>

Teaser: Using `julia --check-bounds=no --threads=1`, I get
```julia
julia> using BenchmarkTools, Trixi

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "structured_3d_dgsem", "elixir_euler_ec.jl"))
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 1456 samples with 1 evaluations.
 Range (min … max):  3.321 ms …   4.301 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.359 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.431 ms ± 178.005 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █ ▆▄▄▄▃▂▃▂▁                                                  
  ███████████████▇▇▇▄█▇▇▆▆▇▇▇▆█▇▇▆▄▇▄▅▁▁▁▄▄▄▄▄▆▅▄▄▄▄▆▄▅▆▁▅▆▆▅ █
  3.32 ms      Histogram: log(frequency) by time      4.16 ms <

 Memory estimate: 416 bytes, allocs estimate: 6.

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "structured_3d_dgsem", "elixir_euler_ec.jl"),
                             volume_flux=flux_shima_etal)
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 3334 samples with 1 evaluations.
 Range (min … max):  1.460 ms …  1.930 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.487 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.497 ms ± 38.342 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄    ▃█     ▂▄                                              
  █▄▁▁▁███▆█▇▆██▇▆██▇██▆▆▅▄▄█▆▅▅▅▅▄▇▁▃▁▃▃▁▅▃▃▃▁▃▃▁▁▁▃▁▄▁▅▄▅▆ █
  1.46 ms      Histogram: log(frequency) by time     1.71 ms <

 Memory estimate: 416 bytes, allocs estimate: 6.

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "structured_3d_dgsem", "elixir_euler_ec.jl"),
                             volume_flux=flux_shima_etal, surface_flux=FluxRotated(flux_shima_etal)) # no direct version available, needs `FluxRotated`
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 2999 samples with 1 evaluations.
 Range (min … max):  1.617 ms …  2.334 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.653 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.665 ms ± 54.161 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▆      █▇    ▁▄▃ ▁▁ ▂▁▁                                    ▁
  █▄▄▄▃▃▆████████████████▇▇▇█▇▇▆▆▁▃▃▄▆▆▄▄▃▃▃▃▃▁▅▁▁▁▃▃▁▃▃▁▁▃▄ █
  1.62 ms      Histogram: log(frequency) by time     1.88 ms <

 Memory estimate: 416 bytes, allocs estimate: 6.
```
on `main` and
```julia
julia> using BenchmarkTools, Trixi

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "structured_3d_dgsem", "elixir_euler_ec.jl"))
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 2368 samples with 1 evaluations.
 Range (min … max):  2.088 ms …  2.514 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     2.097 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   2.109 ms ± 32.061 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄▃▆█▄        ▁▂▅     ▁    ▁▁                                
  █████▅▁▃▄▅▅▅▆███▇▆▆▆███▇▇███▆▅▅▄▆▅▄▄▅▇▇▆▃▃▄▁▅▄▃▁▁▆▇▅▃▁▁▃▁▅ █
  290 ms       Histogram: log(frequency) by time     2.23 ms <

 Memory estimate: 416 bytes, allocs estimate: 6.

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "structured_3d_dgsem", "elixir_euler_ec.jl"),
                             volume_flux=flux_shima_etal)
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 3903 samples with 1 evaluations.
 Range (min … max):  1.244 ms …  2.337 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.268 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.279 ms ± 49.904 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▄     ▄█▁      ▂▃       ▁                                  ▁
  ██▄▄▄▃████▇▇▇▇▇██▇▇▅▇█▇███▇▅▅▇▆▅▆█▆▆▆▄▄▅▄▄▆▃▄▄▅▁▁▃▄▄▁▄▃▁▃▄ █
  1.24 ms      Histogram: log(frequency) by time     1.44 ms <

 Memory estimate: 416 bytes, allocs estimate: 6.

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "structured_3d_dgsem", "elixir_euler_ec.jl"),
                             volume_flux=flux_shima_etal, surface_flux=flux_shima_etal) # No need for `FluxRotated` anymore
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 4035 samples with 1 evaluations.
 Range (min … max):  1.187 ms …  2.787 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.211 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.237 ms ± 82.842 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃  █   ▄▄ ▂▂▂   ▁                                          ▁
  █▃▄███▇███████▇███▇▇▇█▇▆▆▇▇▇▆▆▆▆▆▅▃▆▆▃▅▅▃▅▄▃▄▄▃▄▅▃▁▁▄▁▃▁▄▃ █
  1.19 ms      Histogram: log(frequency) by time     1.57 ms <

 Memory estimate: 416 bytes, allocs estimate: 6.
```
from this PR. 

</details>

Benchmarks on Rocinante are running. I will add them later when they are finished (in a few hours?).