@ranocha 2021-06-30
The curvilinear meshes are still more expensive than the Cartesian `TreeMesh`, of course.
```julia
julia> using BenchmarkTools, Trixi

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "tree_3d_dgsem", "elixir_euler_ec.jl"))
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 1569 samples with 1 evaluations.
 Range (min … max):  2.988 ms …   5.868 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.081 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.182 ms ± 287.739 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▃█▅▅▅▄▄▄▂▃▁▁▁▁▁▁                                             
  ████████████████████▇▇█▇▅▇▆▆▆▅▅▆▆▆▄▁▆▁▆▅▆▆▅▆▆▆▅▆▅▅▅▅▅▅▁▄▁▄▅ █
  2.99 ms      Histogram: log(frequency) by time       4.3 ms <

 Memory estimate: 384 bytes, allocs estimate: 6.

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "tree_3d_dgsem", "elixir_euler_ec.jl"),
                   mesh=StructuredMesh((8, 8, 8), float.(coordinates_min), float.(coordinates_max)))
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 1151 samples with 1 evaluations.
 Range (min … max):  3.634 ms …   5.902 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     4.320 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   4.330 ms ± 260.993 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

                     ▁▁▁▂▆▅▆▄█▆▄▂▅▄▃▅▇▄▆▂▂                     
  ▃▂▂▁▄▄▃▂▄▅▅▄▆▆██▇███████████████████████▇▇▆▆▇▅▄▅▃▃▃▃▄▃▃▃▁▂▃ ▅
  3.63 ms         Histogram: frequency by time         530 ms <

 Memory estimate: 368 bytes, allocs estimate: 6.

julia> begin # RHS
           redirect_stdout(devnull) do
               trixi_include(joinpath(examples_dir(), "tree_3d_dgsem", "elixir_euler_ec.jl"),
                   mesh=P4estMesh((1, 1, 1), polydeg=3, initial_refinement_level=3; coordinates_min, coordinates_max))
           end
           u_ode = copy(sol.u[end])
           du_ode = similar(u_ode)
           @benchmark Trixi.rhs!($du_ode, $u_ode, $semi, $0.0)
       end
BechmarkTools.Trial: 765 samples with 1 evaluations.
 Range (min … max):  6.338 ms …   9.456 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     6.392 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   6.535 ms ± 332.859 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▇▅▃▃▃▂ ▂ ▁▁▁                                                
  █████████▇████▇▆▆▇▇▄█▆▆▇▅▇▇▅█▄▆▄▄▅▄▆▄▁▁▁▄▄▄▁▇▆▄▄▆▅▁▁▁▄▁▁▄▄▄ ▇
  6.34 ms      Histogram: log(frequency) by time      7.92 ms <

 Memory estimate: 736 bytes, allocs estimate: 11.
```
Note that the `P4estMesh` has quite some overhead due to issues such as #642 and #628.