@andrewwinters5000 2021-06-30
Thanks for testing this @ranocha ! Having the volume flux able to accept the pre-averaged contravariant vectors directly make things cleaner and actually somewhat similar to how FLUXO does it ;)